Moved `Position` to its own module and add `Splittable` instance.
